+++ controller/app_tree.py
@@ -3,6 +3,10 @@
 from dataclasses import dataclass
 
 from django.utils.safestring import mark_safe
+from druncschema.controller_pb2 import Status
+
+from interfaces.controller_interface import get_controller_status, get_detectors
+from interfaces.process_manager_interface import get_hostnames
 
 
 @dataclass
@@ -42,3 +46,36 @@ class AppTree:
             table_data.extend(child.to_list(indent + "⋅" + "&nbsp;" * 8))
 
         return table_data
+
+
+def get_app_tree(
+    user: str,
+    status: Status | None = None,
+    hostnames: dict[str, str] | None = None,
+    detectors: dict[str, str] | None = None,
+) -> AppTree:
+    """Get the application tree for the controller.
+
+    It recursively gets the tree of applications and their children.
+
+    Args:
+        user: The user to get the tree for.
+        status: The status to get the tree for. If None, the root controller status is
+            used as the starting point.
+        hostnames: The hostnames of the applications. If None, the hostnames are
+            retrieved from the process manager.
+        detectors: The detectors reported by the controller for each application.
+
+    Returns:
+        The application tree as a AppType object.
+    """
+    status = status or get_controller_status()
+    hostnames = hostnames or get_hostnames(user)
+    detectors = detectors or get_detectors()
+
+    return AppTree(
+        status.name,
+        [get_app_tree(user, app, hostnames, detectors) for app in status.children],
+        hostnames.get(status.name, "unknown"),
+        detectors.get(status.name, ""),
+    )
