+++ interfaces/__init__.py
@@ -1,0 +1,1 @@
+"""Contains the interfaces for all of DRUNC endpoints."""
