+++ interfaces/controller_interface.py
@@ -8,13 +8,9 @@ from drunc.connectivity_service.client import ConnectivityServiceClient
 from drunc.controller.controller_driver import ControllerDriver
 from drunc.utils.grpc_utils import pack_to_any
 from drunc.utils.shell_utils import create_dummy_token_from_uname
-from druncschema.controller_pb2 import Argument, FSMCommand, FSMResponseFlag, Status
+from druncschema.controller_pb2 import Argument, FSMCommand, FSMResponseFlag
 from druncschema.generic_pb2 import bool_msg, float_msg, int_msg, string_msg
 from druncschema.request_response_pb2 import Description
-
-from process_manager.process_manager_interface import get_hostnames
-
-from .app_tree import AppTree
 
 MSG_TYPE = {
     Argument.Type.INT: int_msg,
@@ -132,39 +128,6 @@ def process_arguments(  # type: ignore[misc]
         processed[arg.name] = pack_to_any(MSG_TYPE[arg.type](value=arguments[arg.name]))
 
     return processed
-
-
-def get_app_tree(
-    user: str,
-    status: Status | None = None,
-    hostnames: dict[str, str] | None = None,
-    detectors: dict[str, str] | None = None,
-) -> AppTree:
-    """Get the application tree for the controller.
-
-    It recursively gets the tree of applications and their children.
-
-    Args:
-        user: The user to get the tree for.
-        status: The status to get the tree for. If None, the root controller status is
-            used as the starting point.
-        hostnames: The hostnames of the applications. If None, the hostnames are
-            retrieved from the process manager.
-        detectors: The detectors reported by the controller for each application.
-
-    Returns:
-        The application tree as a AppType object.
-    """
-    status = status or get_controller_status()
-    hostnames = hostnames or get_hostnames(user)
-    detectors = detectors or get_detectors()
-
-    return AppTree(
-        status.name,
-        [get_app_tree(user, app, hostnames, detectors) for app in status.children],
-        hostnames.get(status.name, "unknown"),
-        detectors.get(status.name, ""),
-    )
 
 
 def get_detectors(description: Description | None = None) -> dict[str, str]:
